+++ src/test/java/de/codescape/jira/plugins/scrumpoker/action/ShowErrorLogActionTest.java
@@ -1,0 +1,7 @@
+package de.codescape.jira.plugins.scrumpoker.action;
+
+public class ShowErrorLogActionTest {
+
+    // TODO add tests
+
+}