+++ docs/changelog.md
@@ -43,6 +43,7 @@ Click on the version in the following list to see all changes since the previous
 * feat: add upgrade task with default value for permission check (#68)
 * feat: buttons to confirm estimation only displayed if permitted (#68)
 * docs: update online documentation for Scrum Poker configuration (#75)
+* feat: add error log page to the plugin administration to display errors
 
 ### [4.1.0] - 2019-07-16